break up searchd into modules
