+++ codex/librarian/search/__init__.py
@@ -1,0 +1,1 @@
+"""Search Indexer Thread."""
